remove old header file
