+++ src/pages/Problems.tsx
@@ -1,4 +1,4 @@
-import Header from '../components/header.tsx'
+import Header from '../components/Header.tsx'
 import Footer from '../components/Footer.tsx'
 
 function Problems() {
